session20/requirements.txt
diffusers==0.2.4
gradio==3.49.0
ipython==8.15.0
matplotlib==3.8.0
numpy==1.26.1
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.25.1